added update function

+++ frontend/src/components/NoteDetails.js
@@ -1,20 +1,25 @@
-import React from "react";
+import React, { useState } from "react";
 import { useNotesContext } from "../hooks/useNotesContext";
 
 //date formatting
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
-const NoteDetails = ({ SetButton, setHeading, note }) => {
+const NoteDetails = ({ SetButton, SetSingleNote, setHeading, note }) => {
 
   const { dispatch } = useNotesContext()
+  const [isDeleting, setIsDeleting] = useState(false);
 
   const handleEdit = () => {
     setHeading("Update a note");
     SetButton("Update Note");
+    SetSingleNote(note);
     document.querySelector(".popup-box").classList.add("active");
   };
 
   const handleDelete = async () => {
+    if (isDeleting) return;
+
+    setIsDeleting(true);
     const response = await fetch("/api/notes/" + note._id, {
       method: "DELETE"
     })
@@ -24,6 +29,7 @@ const NoteDetails = ({ SetButton, setHeading, note }) => {
     if(response.ok){
       dispatch({type: "DELETE_NOTE", payload: json})
     }
+    setIsDeleting(false);
   };
 
   return (
@@ -38,7 +44,7 @@ const NoteDetails = ({ SetButton, setHeading, note }) => {
         </span>
         <div className="icons">
           <span className="material-symbols-outlined" onClick={handleEdit}>edit</span>
-          <span className="material-symbols-outlined" onClick={handleDelete}>delete</span>
+          <span className="material-symbols-outlined" onClick={handleDelete} disabled={true}>delete</span>
         </div>
       </div>
     </li>
